tried to implement with getWeather() and getWeatherAndForecastgetWeatherAndForecast()
